+++ src/Service/AuthenticationService.php
@@ -7,34 +7,18 @@ declare(strict_types=1);
 
 namespace WizaplaceFrontBundle\Service;
 
-use Symfony\Component\HttpFoundation\Session\SessionInterface;
-use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
+use Symfony\Component\Security\Core\Authentication\AuthenticationManagerInterface;
 use Symfony\Component\Security\Core\Authentication\Token\UsernamePasswordToken;
-use Wizaplace\SDK\ApiClient;
 use Wizaplace\SDK\Authentication\BadCredentials;
-use Wizaplace\SDK\User\UserService;
-use WizaplaceFrontBundle\Security\User;
 
 class AuthenticationService
 {
-    /** @var ApiClient */
-    private $apiClient;
+    /** @var AuthenticationManagerInterface */
+    private $authManager;
 
-    /** @var TokenStorageInterface */
-    private $tokenStorage;
-
-    /** @var SessionInterface */
-    private $session;
-
-    /** @var UserService */
-    private $userService;
-
-    public function __construct(ApiClient $apiClient, TokenStorageInterface $tokenStorage, SessionInterface $session, UserService $userService)
+    public function __construct(AuthenticationManagerInterface $authManager)
     {
-        $this->apiClient = $apiClient;
-        $this->tokenStorage = $tokenStorage;
-        $this->session = $session;
-        $this->userService = $userService;
+        $this->authManager = $authManager;
     }
 
     /**
@@ -46,10 +30,7 @@ class AuthenticationService
      */
     public function authenticate(string $email, string $password): void
     {
-        $apiKey = $this->apiClient->authenticate($email, $password);
-        $user = new User($apiKey, $this->userService->getProfileFromId($apiKey->getId()));
-        $token = new UsernamePasswordToken($user, null, 'register', $user->getRoles());
-        $this->tokenStorage->setToken($token);
-        $this->session->start(); // Ensure the session exists
+        $token = new UsernamePasswordToken($email, $password, 'main', []);
+        $this->authManager->authenticate($token);
     }
 }